Replace _sequence with _sequence_new
